complete categoryModel to create and delete categories table
