+++ server/models/connection.json
@@ -1,3 +1,3 @@
 {
-  "PG_URI": "postgres://pvjywktf:3lB6lLLkomtvfdffLQzN32BlCM13ossw@fanny.db.elephantsql.com/pvjywktf"
+  "PG_URI": "postgres://ysjvupiw:GS42FNosvS7Nc-XfwfEicKmocUjjpLGg@fanny.db.elephantsql.com/ysjvupiw"
 }
