+++ server/models/setup.js
@@ -1,5 +1,4 @@
 const { Pool } = require('pg');
-
 const { PG_URI } = require('./connection.json');
 
 const pool = new Pool({
@@ -7,137 +6,4 @@ const pool = new Pool({
   max: 5,
 });
 
-async function initDatabase() {
-  await pool.query(`
-    CREATE TABLE categories(
-      name VARCHAR NOT NULL PRIMARY KEY
-    );
-  `, []).then(result => {
-    console.log(result);
-  });
-
-  await pool.query(`
-    CREATE TABLE recommendations(
-      id SERIAL NOT NULL PRIMARY KEY,
-      title VARCHAR NOT NULL,
-      body VARCHAR NOT NULL,
-      date TIMESTAMP NOT NULL,
-      category VARCHAR,
-      rating INT NOT NULL,
-      FOREIGN KEY (category) REFERENCES categories(name)
-    );
-  `, []).then(result => {
-    console.log(result);
-  });
-
-  await pool.query(`
-    CREATE TABLE users(
-      username VARCHAR NOT NULL PRIMARY KEY,
-      name VARCHAR NOT NULL,
-      email VARCHAR NOT NULL,
-      passhash VARCHAR NOT NULL,
-      last_login_ip VARCHAR NOT NULL,
-      last_login_date TIMESTAMP NOT NULL
-    );
-  `, []).then(result => {
-    console.log(result);
-  });
-
-  await pool.query(`
-    CREATE TABLE user_recommendations(
-      id SERIAL NOT NULL PRIMARY KEY,
-      username VARCHAR NOT NULL,
-      recommendation_id INT NOT NULL,
-      FOREIGN KEY (username) REFERENCES users(username),
-      FOREIGN KEY (recommendation_id) REFERENCES recommendations(id)
-    );
-  `, []).then(result => {
-    console.log(result);
-  });
-
-  await pool.query(`
-    CREATE TABLE user_follows(
-      id SERIAL NOT NULL PRIMARY KEY,
-      username VARCHAR NOT NULL,
-      followed_username VARCHAR NOT NULL,
-      FOREIGN KEY (followed_username) REFERENCES users(username)
-    );
-  `, []).then(result => {
-    console.log(result);
-  });
-
-  await pool.query(`
-    CREATE TABLE sessions(
-      id SERIAL NOT NULL PRIMARY KEY,
-      username VARCHAR NOT NULL,
-      expires TIMESTAMP NOT NULL,
-      FOREIGN KEY (username) REFERENCES users(username)
-    );
-  `, []).then(result => {
-    console.log(result);
-  });
-}
-
-async function destroyDatabase() {
-  try {
-    await pool.query(`
-      DROP TABLE sessions;
-    `, []).then(result => {
-      console.log(result);
-    });
-  } catch (_) {
-    // its okay if it doesn't exist, calm down
-  }
-
-  try {
-    await pool.query(`
-    DROP TABLE user_follows;
-  `, []).then(result => {
-      console.log(result);
-    });
-  } catch (_) {
-    // its okay if it doesn't exist, calm down
-  }
-
-  try {
-    await pool.query(`
-    DROP TABLE user_recommendations;
-  `, []).then(result => {
-      console.log(result);
-    });
-  } catch (_) {
-    // its okay if it doesn't exist, calm down
-  }
-
-  try {
-    await pool.query(`
-    DROP TABLE users;
-  `, []).then(result => {
-      console.log(result);
-    });
-  } catch (_) {
-    // its okay if it doesn't exist, calm down
-  }
-
-  try {
-    await pool.query(`
-    DROP TABLE recommendations;
-  `, []).then(result => {
-      console.log(result);
-    });
-  } catch (_) {
-    // its okay if it doesn't exist, calm down
-  }
-
-  try {
-    await pool.query(`
-    DROP TABLE categories;
-  `, []).then(result => {
-      console.log(result);
-    });
-  } catch (_) {
-    // its okay if it doesn't exist, calm down
-  }
-}
-
-module.exports = { initDatabase, destroyDatabase, pool };
+module.exports = { pool };
